added clangd, fix problems in code

+++ cpu/idt.c
@@ -1,10 +1,11 @@
+#include <stddef.h>
 #include "idt.h"
 #include "screen.h"
 
 idtEntry idt[IDT_ENTRIES];
 idtRegister idtReg;
 
-void setIdtGate(uint16_t n, uint32_t address) {
+void setIdtGate(uint16_t n, size_t address) {
     idt[n].offsetL = address & 0xffff;
     idt[n].selector = GDT_CS_OFFSET;
     idt[n].zero = 0;
@@ -20,7 +21,7 @@ void setIdtGate(uint16_t n, uint32_t address) {
 }
 
 void initIdt() {
-    idtReg.base = (uint32_t)&idt;
+    idtReg.base = (size_t)&idt;
     idtReg.limit = sizeof(idtEntry) * IDT_ENTRIES - 1; // -1 because sizes start at 1
     __asm__ __volatile__("lidtl (%0)" : : "r" (&idtReg));
 }
